Selbonyka patch 1 (#1)
* Create 1MistakeFixed

Since this is written in Python, we don't need to have the semicolons at the end of each lines

* Create 2ndMistakeFixed

Variables were names differently which prevented proper execution of the code. As well they were named not according to PEP8.

* Create 3rdMistakeFixed

Last mistake found in the code - wrong spelling of the word "second.